Map integration with new page on pin click
I've been able to integrate the map api and manage to make the pins load a second onsen page on click.

The pin clicking doesn't function very well in the browser but works fine on my iPhone. I've also only manage to get everything working by having the new js in the html document. As soon as I try copying everything over to the js file it breaks and I can't work out a way to do this. The next step is for me to work out how I can make your 'createPara' and button functions appear in the output id on page 2. As soon as I delete the section containing the 'id= output' from page 1, it breaks the code, even though I've included it under page 2.